Some suggested tweaks
This adds padding to the <h2> at the top. And makes the first
"MDN Web Docs" into a link and then styles the link specially
because it is already on a blue background. And it changes
"Get in contact" to "Get in touch" because the old wording sounds
awkward to my ear.